Refs #98 - Moved tests outside django-downloadview package, i.e. in tests/ folder at project root.
