+++ tox.ini
@@ -16,8 +16,8 @@ deps =
 commands =
     pip install -e .
     pip install -e demo
-    demo test --cover-package=django_downloadview {posargs: django_downloadview demoproject}
-    cover erase
+    demo test --cover-package=django_downloadview --cover-package=demoproject {posargs: tests demoproject}
+    coverage erase
     pip freeze
 
 [testenv:flake8]
